// portal_spa/src/models/environment.ts
export type Environment = {
  id: number;
  environment: string;
  name_en: string;
  name_nl: string;
  description_en: string;
  description_nl: string;
  url: string;
  scheme_id: string;
  timestamp_server: string;
  keyshare_server: string;
  keyshare_website: string;
  keyshare_attribute: string;
  contact_website: string;
};